small fixes + measure time in dmatrix shuffle

+++ src/dft_ground_state.cpp
@@ -31,9 +31,9 @@ double DFT_ground_state::energy_enuc()
     double enuc = 0.0;
     if (parameters_.unit_cell()->full_potential())
     {
-        for (int ialoc = 0; ialoc < (int)parameters_.unit_cell()->spl_num_atoms().local_size(); ialoc++)
+        for (size_t ialoc = 0; ialoc < parameters_.unit_cell()->spl_num_atoms().local_size(); ialoc++)
         {
-            int ia = parameters_.unit_cell()->spl_num_atoms(ialoc);
+            int ia = parameters_.unit_cell()->spl_num_atoms((int)ialoc);
             int zn = parameters_.unit_cell()->atom(ia)->type()->zn();
             enuc -= 0.5 * zn * potential_->vh_el(ia) * y00;
         }
